+++ lib/rubocop/cop/style/format_string.rb
@@ -48,6 +48,10 @@ module RuboCop
         }
         PATTERN
 
+        def_node_matcher :variable_argument?, <<-PATTERN
+          (send {str dstr} :% {send_type? lvar_type?})
+        PATTERN
+
         def on_send(node)
           formatter(node) do |selector|
             detected_style = selector == :% ? :percent : selector
@@ -70,10 +74,10 @@ module RuboCop
         end
 
         def autocorrect(node)
-          lambda do |corrector|
-            detected_method = node.method_name
+          return if variable_argument?(node)
 
-            case detected_method
+          lambda do |corrector|
+            case node.method_name
             when :%
               autocorrect_from_percent(corrector, node)
             when :format, :sprintf